requirements.txt
pydantic
anthropic>=0.18.0
python-dotenv
google-genai
# stuff for low quality web app made by llms :(
fastapi